Authentifications des routes

+++ src/pages/catways.js
@@ -11,10 +11,26 @@ function Catways(params) {
 
     const fetchDataCatways = async () => {
         try {
-            const response = await fetch('https://express-api-port-plaisance-russell.onrender.com/api/catways');
+            // Récupérer le token JWT du sessionStorage
+            const token = sessionStorage.getItem('token');
+    
+            // Vérifier si le token JWT existe
+            if (!token) {
+                // Rediriger vers la page de connexion s'il n'y a pas de token
+                window.location.href = '/AuthentificationCatwaysD';
+                return; // Arrêter l'exécution de la fonction
+            }
+    
+            const response = await fetch('https://express-api-port-plaisance-russell.onrender.com/api/catways', {
+                headers: {
+                    'Authorization': `Bearer ${token}`
+                }
+            });
+    
             if (!response.ok) {
                 throw new Error('Error fetching catways');
             }
+    
             const data = await response.json();
             setCatways(data);
         } catch (error) {
@@ -22,16 +38,32 @@ function Catways(params) {
         }
     };
 
-    const deleteCatway = async (catwayId) => {
-        try {
-            await fetch(`https://express-api-port-plaisance-russell.onrender.com/api/catway/${catwayId}`, {
-                method: 'DELETE'
-            });
-            fetchDataCatways();
-        } catch (error) {
-            console.error('Error deleting catway:', error);
-        }
-    };
+        const deleteCatway = async (catwayId) => {
+                try {
+                // Récupérer le token JWT du sessionStorage
+                const token = sessionStorage.getItem('token');
+                
+                // Vérifier si le token JWT existe
+                if (!token) {
+                        // Rediriger vers la page de connexion s'il n'y a pas de token
+                        window.location.href = '/AuthentificationCatwaysD';
+                        return; // Arrêter l'exécution de la fonction
+                }
+        
+                // Envoyer la requête DELETE avec le token JWT dans le header Authorization
+                await fetch(`https://express-api-port-plaisance-russell.onrender.com/api/catway/${catwayId}`, {
+                        method: 'DELETE',
+                        headers: {
+                        'Authorization': `Bearer ${token}`
+                        }
+                });
+        
+                // Mettre à jour les données après la suppression du catway
+                fetchDataCatways();
+                } catch (error) {
+                console.error('Error deleting catway:', error);
+                }
+        };
 
         return(
                 <div>
